2095301: enable sslverifystatus on repos if advertized by CP
An upcoming change to Candlepin is the addition of a capability to
advertize the possibility to perform OCSP checking when connecting to
its repositories. If it is possible, set the "sslverifystatus" repo key
to "1", as dnf supports it already.

(cherry picked from commit 9fd6af6d62cb0562ccaf0d431df22f63077c8826)