2092014: Do not use injections in rhsm
* BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2092014
* Card ID: ENT-240

Using injections is proven to be... unreliable. They work good enough
when used in subscription-manager, but when they are pulled out into
other modules (e.g. rhsm), they start to break.

The problem with current state was following: when rhsm was used as
library by some other RPM (e.g. virt-who), it tried to import
subscription_manager as a package. If it was not installed on the
system, everything worked. But if it was installed on the system, the
`import` statement succeeded, and even `inj.PROGRESS_MESSAGES` was
usable.

But injections have to be initialized, specifically by calling function
`init_dep_injection()` from `injectioninit.py`. That is not the case
when sub-man is used as a library. So in this situation rhsm crashed
because subscription-manager's injections didn't have this specific
injection registered, which resulted in KeyError being raised.

As a result of that, we're not going to use injections at all. Instead,
we can use module variable as a singleton, which is much easier and more
reliable solution.