ci: drop stylish job from Jenkins
The style checks are now done using a GitHub workflow, so we do not need
this step anymore.

The only regression is the lack of rpmlint test; this is because the
current config for it fails with rpmlint >= 2, so it will be fixed in
followup commits when the test is not run.