build: switch rpmlint check to rpmlint 2.x
rpmlint 2.x changes almost radically the command line option, forcing
to switch to the new format in case we do not want to do "autodetection"
(which is not a good idea in the long term, anyway).

This also needs to adapt the configuration with the filters/badness,
removing filters that do not apply anymore: they either are no more
issues in the spec file (dnf plugin & tmpfiles.d paths), or rpmlint does
not complain about them (the zypper plugin path).