build: fix extraction of cockpit messages
It seems like the "JavaScript" language of gettext is not able to handle
JS code with React bits inside.

Instead, use the same approach used by the cockpit project (and already
in use in the separate subscription-manager-cockpit repository): use "C"
for the JS sources, adding the minimal keywords needed to extract all
the messages currently in the plugin.