cloud-what: switch MiniHostCollector to dmidecode(1)
The python-dmidecode module has currently not received work upstream in
the last 7 years, and it starts to fails in some situations (e.g. Kernel
Lockdown, when /dev/mem cannot be read).

As way around that, switch the MiniHostCollector facts collector to
invoke the dmidecode(1) tool instead to get some facts from known DMI
strings.

The general result of the detection should still be the same, as the
fetched strings are used (or potentially used) during the cloud
detection by the providers. Also, this fact detector is used only when
subscription-manager is not installed, so the impact should be minimal.

Adapt the buildsystem & packaging bits to require dmidecode for
cloud-what rather than the Python module dmidecode.