facts: add dmidecode parser and facts collector
Create a new parser for parsing the output of the dmidecode(1) tool,
collecting all the data from it. Add a test for it with existing outputs
to check, as dmidecode information can be read only as root, and they
would change on each system. The data-based tests check that the parser
works fine, and extracts at least the system UUID.
The test data have sensible details (UUIDs, serial numbers, product
names, and part numbers) replaced with placeholders; this also helps
testing that the parsed output contains the values we expect.

In addition to that, add a facts collector that uses this new parser;
to make sure the facts resemble somehow what the python-dmidecode-based
one returns, do few quirks to adapt the data read from dmidecode(1).
Add a test for this new collector, using the aforementioned test data,
and checking that at least some subtags of "dmi" facts are collected.