connection: improve the internal UnknownContentException
Make it carry the content type and the actual content of the reply from
the server. This way it will be possible to provide a better diagnostic
for it.

Adapt the various places that raise it to provide all the attributes
they know about.

(cherry picked from commit 7407df9709a5b1bdc8e7c5e57d6920daa5f9a4f7)