Improve formatting of UnknownContentException
In case of unknown/unhandled content reply from the server, format it
properly by printing the HTTP error code & its description, and the
actual content if available. Even this may seem verbose on terminal, at
least it allows the user to see what is the actual failure without the
need to look at the log file or enable the debugging of the HTTP
traffic.