Run multiple Python versions in CI
Previously, everything was run in a Fedora container. This new version
splits rpmlint from the rest, which gives us a few advantages.

The first is speed. We don't have to spin up several Fedora containers
now, which makes the results visible in 30-40 seconds, instead of a
minute or two.

The second is pip caching, which is supported by actions/setup-python.
This also contributes to the speedup.

The last thing is the fact that we don't have to do anything to make
all the different Python versions work. It could be difficult to handle
installation and setup of all the Python versions.

One disadvantage is that now we rely on Ubuntu system and not RHEL-based
system. I don't think this makes it less reliable though, as it is only
linting and no libraries we use have to be installed.