ENT-5317: New way to test DBus methods, starting with Config
* Card ID: ENT-5317

We've seen some strange issues when the full test suite has been run
several times in a row. A process responsible for creating DBus session
was not being destroyed *every time* -- when it wasn't, the next DBus
test would try to claim the same DBus address again. That would raise an
exception in child thread (that was not handled), pytest would get stuck
until the timeout happened; the main thread was still asleep, resulting
in INTERNALERRORs.

This new approach addresses this in several ways:

- DBus server is created for whole DBus class, not for every one of the
  tests. That means that even if the probability of hanging up stayed
  here, the chance of that happening decreases by an order of magnitude.

- The tests aren't using DBus to pass data around now. While the server
  is needed to instantiate the DBus classes, we can use Python magic to
  access original functions that were decorated and call them instead,
  circumventing all DBus communication.

This has a side effect of speeding up the tests, as both server creation
and DBus transmission added non-zero delay to each of DBus tests.

(Cherry-picked from 2e5dd64)