cockpit: disable the reference branch for sub-man-cockpit
By default, cockpit's run-tests tries to get the list of tests that
changed in the branch being tested; since we run the test suite of
subscription-manager-cockpit from the local clone of it, that repository
will definitely not contain subscription-manager branches.

As a possible solution, disable the reference branch for run-tests
(using an empty string): we are not changing sub-man-cockpit tests from
this repository, so there are definitely no changes in its tests.