Raise a new CertificateLoadingError on X.509 loading failures
Change the C extension to raise a new OpenSSLCertificateLoadingError
exception in case the loading of a X.509 certificate fails, rather
than returning a null object: this way it is possible to properly
propagate also a string with the OpenSSL error. This new exception only
carries the bits from OpenSSL, so keep it mostly internal; add a
CertificateLoadingError representing the public exception for this kind
of sitation.

Because of this:
- drop the check in _CertFactory._read_x509, as the x509 object is
  always non-None
- wrap the certificate loading in _CertFactory.create_from_file, and
  _CertFactory.create_from_pem to raise CertificateLoadingError
- catch the new CertificateLoadingError in CliCommand.main(), doing the
  same job of CertificateException

There should be no behaviour difference in subscription-manager.

(cherry picked from commit 05f20b9713e4bb629243da947da9d53226ba9877)