[1.29.30] 2142611: Clear progress messages properly
* BZ: 2142611
* Card ID: ENT-5468

Multi-byte progress messages, such as Japanese, were not being cleared
correctly. The number of characters was lower than the number of bytes
written to the console, so we weren't clearing enough.

Escape code 0K should handle this properly, as it will delete all the
content on current line.

Cherry picked from 62652ba.