Fix failures of D-Bus' Register tests
Our test that checks if it is possible to connect to newly created
socket started failing with 'Permission denied' when the code tried to
use it. Pytest cannot write into '/run', moving it to '/tmp' solves the
issue.

Additionally, as a result of relatively recent D-Bus implementation
changes, 'tmpdir' and 'dir' behave the same and create regular socket,
not an abstract socket. Connecting with "\0" prefix (which signifies the
abstract one) now fails and thus has been removed from the tests.

(Cherry-picked from bae7ce3)