utils: import pkg_resources only when needed
Merely importing "pkg_resources" will perform its initialization, which
is not exactly lightweight. "pkg_resources" is used to get the version
of subscription-manager as installed, in case there is no valid version
in the internal "version" module, which usually happens in
developer-only setups.

Since the build system takes care of filling "pkg_version" in the
"version" module, a distro build of subscription-manager does not need
to query "pkg_resources" for the version. Hence, import "pkg_resources"
only when we actually need to use it for getting the version.