libdnf: fix return value of findProductId()
findProductId() returns an int, either -1 or 1 depending on whether it
was possible to find the product ID in the specified certificate.
The problem with this kind of return value is that's very easy to
mistake for a boolean, and indeed installProductId() has a boolean-like
check that it is always true, no matter the return value of
findProductId().

As a fix, change the return type of findProductId() to gboolean, with
true returned whether the search succeeded:
- reduce the scope of the 'ret_val' variable, so there are less chances
  to use it wrongly
- make 'ret_val' false by default, only changing it to true in the only
  place where it needs to be like that
- adapt the tests accordingly

Reported-by: Matyas Horky