2174297: register: do a simple strip() on environment(s) input
When we prompt the user for multiple environments, in the input string
any space is removed before further internal processing.  Most likely
this was done to minic the non-multi-environment case that strips the
input to ensure it is not empty; sadly this makes more than that,
making it impossible to interactively type environment names that
contain a space, as it will be removed.

Since the result of _prompt_for_environment() is passed to
check_set_environment_names(), which strip()s any environment after
splitting the input string by comma, we can simply strip() the input
string in both multi-environments and non-multi-environments cases: it
will be enough to ensure that it is non-empty.

Add a simple test to check a couple of these whitespace situations,
i.e. in the middle after a comma and at the end.