Fix collector test for aarch64
In test/rhsmlib/facts/test_collector.py, the test
test_get_aarch64_uuid_collection_no_file incorrectly mocked file not
being found. This was not an issue until recently when we started to
run the test on ARM as well as on x64.

This test is not present in newer branches, as they are using dmidecode
data for both of those architectures.