Fix i18n setup on Python 3.10+
gettext.bind_textdomain_codeset has been deprecated since Python 3.8 and
was removed in 3.10. To make the branch valid on new versions of Python
as well (for development reasons), this patch will allow i18n to be
set up correctly. As it does pretty much the same as its locale.*
alternative, there should not be any difference.