Alter import of rhsm.config functions
Because 'initConfig()' and 'in_container()' functions were being
imported directly, instead of as reference, the mocks that tests perform
were not applying to them, causing test suite to fail when run from
inside a container. Future-wise it is better to manage mocks on one
place, instead of patching the functions whenever they are being
imported.

(Adapted from 0b78761)