Revert "ENT-5549: Fix return code handling of CLI"
CLI is subclassed/used not only in the 'subscription-manager' script,
but also by 'rhsm-debug' and 'rct'; since the InvalidCLIOptionError
handling was moved from CLI to 'subscription-manager' script, invalid
options in 'rhsm-debug' and 'rct' cause tracebacks.

The idea of c24da73bd70c8ecd502419e4aa313bc08cf244b2 was to avoid
returning None from CLI.main(), which the code used to do when catching
InvalidCLIOptionError. While the idea was good, we can better fix it
globally still in CLI.main().

This reverts commit c24da73bd70c8ecd502419e4aa313bc08cf244b2.