dbus: don't catch exceptions in DomainSocketServer.run()
Catching all the exceptions in DomainSocketServer.run() has two big
drawbacks in case anything in it fails:
- the error is logged, and then nothing is done with it
- callers of it cannot even know that an issue happened
- even worse: run() returns a string in case of no errors, whereas the
  exception case did not return anything (thus None), causing type
  issues in callers

The only user of DomainSocketServer is RegisterDBusObject, and all the
D-Bus methods already catch all the exceptions that happen during their
execution (turning them into D-Bus exceptions). Hence, it is possible to
safely drop all the exception handling in DomainSocketServer.run(),
letting errors that happen in it properly propagate back as error reply
to D-Bus.