managercli: add a local get_current_owner() helper
Add a local get_current_owner() helper function that mimicks the one
available in newer versions. Backporting the implementation available in
newer version is complicated, and relies on not available bits such as
the owner cache.

This simple method will do, and reduces the differences in a couple of
places with what's in more recent branches.