Update cockpit test lib for Chromium 113 fix
The most recent Chromium 113 does not work any more with sizzle, tests
were hanging as soon as they switched to a frame. The latest cockpit
testlib applied a workaround [1]. As the latest Cockpit library also
moved to PatternFly 5, but that would be prohibitively expensive for
sub-man's stable branch, switch to the (temporary) "274+chromium-sizzle"
branch, which has that workaround backported.

[1] https://github.com/cockpit-project/cockpit/pull/18812