refresh: avoid ExceptionMapper in case of RestlibException
The recent changes to RefreshCommand make use of ExceptionMapper in a
place where it was not used previously. While using ExceptionMapper is
generally a good idea to format errors, in this case the consistency
with the behaviour of a maintenance branch is preferred.

Hence, manually print again the message of RestlibException.

Updates commit 3b3eda26eb058ab20fde1edf25e8a5b5cb9d3200.