ENT-5581: Update messaging around the "container mode"
* Card ID: ENT-5581

The two messages that would be printed in containers were not consistent
and slightly contradictory (is operating x is disabled).

This change unifies them and ensures that the tool name is always
'subscription-manager'. It also updates the CLI message which now better
describes the action necessary to manage the subscriptions via
the host system.