Remove version constraint of pytest
Previously, in f801b69, we had to limit pytest to version <7 because our
tests were using pytest-forked to run D-Bus unit tests and pytest was
crashing on that.

We have since changed the way we test D-Bus by not starting a server for
each test case, and aren't constrained by the pytest regression anymore.