feat: deprecate attach/auto-attach modules
This patch adds a deprecation notice
to the attach/auto-attach command descriptions.
This patch also moves the attach command to the
'Other Modules' section in the sub-man help text
from the 'Primary Modules' section.

The attach command already displays an error
message when attempting to attach subscriptions
with an SCA registered system, so no changes
are required to implement this.

Signed-off-by: Jason Jerome <jajerome@redhat.com>