[ATIX] Handle quoted baseurl
Candlepin 4.4 has started to deliver quote baseurl's. Apparently this
is considered a bug, but there is no telling when it will be fixed on
the candlepin side. In addition unquote() should not harm strings that
are not quoted, so this works either way.

(cherry picked from commit 805d69b459c8fa17939b7e94e3f55d75058dc3b8)