rhsmcertd: refactor timestamp function
Rather than modifying and returning the static buffer of asctime(),
provide a buffer for it on which to copy the resulting timestamp.
Because of this, give the function a new name that indicates what it
does now.

While the static buffer of asctime() is still used, this will make it
possible to change the implementation of this helper function to (also)
not rely on that.

There should be no behaviour changes.