Stop using usermode
Fedora is working on unifying /usr/bin and /usr/sbin:
https://fedoraproject.org/wiki/Changes/Unify_bin_and_sbin
Because of that, we cannot ship anymore /usr/bin/subscription-manager
and /usr/sbin/subscription-manager as different executables.

/usr/bin/subscription-manager is actually provided for usage with
usermode, so non-root users invoking "subscription-manager" are prompted
for the root password. Nowadays most administrators use "sudo" for
running commands, and usermode did not get a lot of adoption.

Hence, simply drop the usage of consolehelper altogether. This makes
subscription-manager compatible with Fedora 42+, and in general drops a
very low used feature.

Card ID: CCT-1106