build: make it possible to customize the 'sbin' install dir
The build system hardcoded 'sbin' as location where to install the
'subscription-manager' executable. While that is generally correct,
make it possible to customize that location to set it to right one
according to the distribution policies.

Use "-n" when moving the 'subscription-manager' executable to not error
out in case SBIN_DIR is set to be 'bin'.

There is no behaviour change, as the default is 'sbin' as previously
hardcoded.