feat: print "not registered" on status when not registered
In case the system is not registered do not attempt to get the
compliance/entitlements state, as it will not reflect the actual status
of the system (i.e. not registered).

In that case there is also no need to print the syspurpose status, as it
is not relevant on unregistered systems.