spec: Avoid DNF warnings on traditional Fedora hosts and DNF on atomics
It will be good to make it easier to have gratis, self-supported Red Hat
Enterprise Linux OCI containers on Fedora Silverblue and Workstation
through the RHEL Developer Suite subscription [1].  This includes Toolbx
containers created with:
  $ toolbox create --distro rhel --release 9.5

However, when subscription-manager was installed on a Fedora host with
DNF4, the subscription-manager DNF plugin printed some spurious warning
messages that users didn't like because it made them think that
something is wrong.  eg., when the host is unregistered:
  $ sudo dnf update
  Updating Subscription Management repositories.
  Unable to read consumer identity

  This system is not registered with an entitlement server. You can use
      subscription-manager to register.

  Last metadata expiration check: ...

Or, when the host has been registered:
  $ sudo dnf update
  Updating Subscription Management repositories.

  This system is registered with an entitlement server, but is not
      receiving updates. You can use subscription-manager to assign
      subscriptions.

  Last metadata expiration check: ...

These warnings have no relevance in Fedora, because the purpose of using
subscription-manager(8) on a Fedora host is to access RHEL content in
OCI containers.  There is no need for RHEL content on the Fedora host
itself.

Note that these warnings aren't visible on Fedora hosts with DNF5
because the subscription-manager plugin doesn't work with it.

Secondly, until now, subscription-manager had %{_bindir}/dnf-3 and
%{_bindir}/dnf4 in its dependency chain.  This meant that installing
subscription-manager on Fedora Silverblue pulled them in, even though
those executables shouldn't be present on Silverblue because they don't
work there.

Both these problems can be addressed by disabling the DNF and libdnf
plugins on Fedora.

[1] https://developers.redhat.com/blog/2016/03/31/no-cost-rhel-developer-subscription-now-available

https://bugzilla.redhat.com/show_bug.cgi?id=2246833
https://bugzilla.redhat.com/show_bug.cgi?id=2262228