spec: Require subpackages from the same build as the base package
While the Fedora packaging guidelines [1] don't say anything about a
base package requiring its subpackages, it's highly unusual to mix and
match RPMs originating from different builds when they share the same
sources.  eg., a downstream patch or build flag may be applied to one of
those RPMs but not applied to another.

Unless there's a very good reason not to, it's better to stick to
established idioms and ensure that RPMs with shared sources are always
used with their siblings from the same build.

Fallout from the following commits:
  * f445b6486a962d12185a5afe69e768d0a605e175
  * 8dddb545597e31a45304bec0d33e89e52a8fa705
  * af28bb91e93f33420e50ac48724739af9782b5af

[1] https://docs.fedoraproject.org/en-US/packaging-guidelines/