spec: Remove a condition that's always false
The DNF plugin is never built on Fedora.  Therefore, it's pointless to
check for Fedora when building it.

In the past, the DNF plugin used to be built on Fedora 29 to 32, and the
libdnf plugin used to be built on Fedora 33 onwards.  That's why it made
sense to check for Fedora for the DNF plugin.  However, that's no longer
the case.

Fallout from a0468a45999bc9cf988a59470219f4da73f0dfaa