spec: Make the Fedora build flags more explicit than non-RHEL
A lot of different downstream distributions are supported by
subscription-manager.spec.  For example, Fedora, RHEL and the SUSE
family.  Representing Fedora as merely non-RHEL can lead to surprises
when dealing with RHEL clones or other Fedora derivatives, or when
refactoring the code.

It's better to be explicit about Fedora because the build flags were
carefully and consciously chosen for it, and any inadvertent change
would be harmful.