fix: Handle improperly closed connections
* Card ID: CCT-1261
* Card ID: RHEL-83016

Sometimes, the TLS connection wouldn't get closed properly (various
implementations of various TLS versions handle things differently), and
the socket unwrap would raise BrokenPipeError.

This patch ensures that all exceptions raised by TLS are logged and
ignored -- we're closing the connection, there's no harm in pretending
the issue didn't happen.

At the same time, we're wrapping the TLS connection close() in
try/except as well. While we haven't seen anything that would suggest
this code is faulty as well, it's better to be safe, and log that
something unexpected happened.