fix: Enable content during auto-registration v2
* Card ID: CCT-1386
* Card ID: RHEL-90416
* The content was not enabled during auto-registration on public clouds,
but only consumer certificate was generated after registration in
certain scenario. It was caused by fact that token returned during
auto-registration (v2) could be returned with tokenType equal to
"CP-Cloud-Registration" (not only "CP-Anonymous-Cloud-Registration").
When `manage_repos` is equal to "1", then we should install SCA
entitlement certificate and generate redhat.repo in this case. When
`manage_repos` is equal to "0", then content will be still consumed from
RHUI on public cloud.

Signed-off-by: Jiri Hnidek <jhnidek@redhat.com>