fix: [1.29] Enable content during auto-registration v2
* Backport to 1.29 branch
* Original PR: #3568
  * Original commit: 42eb05b7e2332311d1ed43499a74300832681973
* Card ID: CCT-1386
* Card ID: RHEL-90416
* The content was not enabled during auto-registration on public clouds,
but only consumer certificate was generated after registration in
certain scenario. It was caused by fact that token returned during
auto-registration (v2) could be returned with tokenType equal to
"CP-Cloud-Registration" (not only "CP-Anonymous-Cloud-Registration").
When `manage_repos` is equal to "1", then we should install SCA
entitlement certificate and generate redhat.repo in this case. When
`manage_repos` is equal to "0", then content will be still consumed
from RHUI on public cloud.

Signed-off-by: Jiri Hnidek <jhnidek@redhat.com>